Prompting the user for arguments via `gets` doesn't really make sense for the target audience of gistory. A better strategy is to provide some usage instructions (and return a non-zero exit status) when the required arguments are not provided.
This allows the user to make full use of their shell (tab completion etc) when providing your program with input values.

Additionally, methods that end with a `?` should *test* for some condition in their environment, their arguments or a combination of the two. Their return value should generally be a boolean.

For example:

  def missiles_lauched?
    @missile_silo.empty?
  end

The `supplied?` method should have been  a verb like `get` or `ask_for`.

`get_commits` was defined then immediately called. Generally you define a method for two reasons:

1. To facilitate code re-use by factoring out common functionality
or
2. To provide a meaningful name for a block of code

In this case it looks like `get_commits` was defined simply to provide a scope for some local variables such as `log_data` and `commit_diff_data`. Ruby provides a better way to this: the `begin` keyword. So

  def get_commits
    ... code that builds the commits ...
  end
  commits = get_commits

can be more clearly written as

  commits = begin
    ... code that builds the commits ...
  end

As you can see on line 12 `begin` blocks can be combined with the `rescue` keyword to form basic control structures:

  repo = begin
    Repo.new(repo_path)
  rescue InvalidGitRepositoryError
    abort "Error: #{repo_path} is not a git repo"
  end